Fixed a bug
There was a problem that if "if" is placed at the head of the line but actually the line is a continuation from the previous line, the "stata_linter_correct" command thought that the line should have a curly bracket and searched for it in the following lines. I added a condition so that, if the line with "if" at the beginning is a continuation from the previous line, then the "stata_linter_correct" command does not think that the expression in the line should have a curly bracket.